Add filtering sellerPageList
