+++ 3MCJ/src/Routing/Routing.tsx
@@ -11,7 +11,6 @@ function Routing() {
       <Route path="/login" element={<Login />} />
       <Route path="/home" element={<Home />} />
       <Route path="/" element={<Home />} />
-      <Route path="/sellerPage" element={<SellerPage />}/>
     </Routes>
   );
 }
